requirements.txt
python-dotenv==1.2.2
streamlit
pandas
